requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.8
gunicorn==19.9.0
